Bug fix for failure to start
When the camera is stopped and started, the web browser may
acquire new image when there is none. This threw a runtime
exception that was caught in 'main' and the program exited.
Now the exception is caught and an error code is returned.
Hence the web server needs to check if the return value is
success "\0\0\0\0" or "EMPT".